Establish layout for a unified competition query from the core module
