+++ packages/arena-interface/src/competition/mod.rs
@@ -1,3 +1,4 @@
 pub mod msg;
 pub mod state;
 pub mod stats;
+pub mod types;
